# server/api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Training app for climbers. Sign up, sign in on several devices at once, and\nsign each one out on its own.\n\nEvery failure answers with the same shape: an error object carrying a code,\na human message, and on a validation failure a fields map naming what to\nfix.",
    "title": "Passion",
    "version": "1"
  },
  "basePath": "/",
  "paths": {
    "/api/v1/accounts": {
      "post": {
        "description": "Answers with the account and a token, so the client signs up and signs in\nwith one call.\n\nAn address that is already registered answers invalid_credentials, the same\nas a wrong password. Telling them apart would let anyone ask the API who has\nan account.",
        "tags": [
          "accounts"
        ],
        "summary": "Create an account",
        "operationId": "signUp",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signUpRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/signUpResponse"
          },
          "401": {
            "$ref": "#/responses/invalidCredentials"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": []
      }
    },
    "/api/v1/accounts/me": {
      "get": {
        "description": "Answers with the account the bearer token belongs to.",
        "tags": [
          "accounts"
        ],
        "summary": "Read the signed-in account",
        "operationId": "readAccount",
        "responses": {
          "200": {
            "$ref": "#/responses/accountResponse"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/exercises": {
      "get": {
        "description": "The exercises the app ships and the ones you wrote, sorted by name. Retired\nones are left out.",
        "tags": [
          "exercises"
        ],
        "summary": "List your library",
        "operationId": "listExercises",
        "responses": {
          "200": {
            "$ref": "#/responses/exerciseListResponse"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "description": "Adds one to your library. A field left out or sent as null is not set.",
        "tags": [
          "exercises"
        ],
        "summary": "Create an exercise",
        "operationId": "createExercise",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/exerciseRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/exerciseResponse"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/exercises/{id}": {
      "get": {
        "description": "One the app ships or one of yours. A retired one still reads.",
        "tags": [
          "exercises"
        ],
        "summary": "Read an exercise",
        "operationId": "readExercise",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The exercise's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/exerciseResponse"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "description": "Sets every field of one of your own. A field left out or sent as null is\ncleared. A shipped exercise, or someone else's, answers 404.",
        "tags": [
          "exercises"
        ],
        "summary": "Replace an exercise",
        "operationId": "updateExercise",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/exerciseRequest"
            }
          },
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The exercise's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/exerciseResponse"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/exercises/{id}/retire": {
      "post": {
        "description": "Takes one of your own out of your library. It still reads by id. Retiring it\ntwice keeps the first date. A shipped exercise, or someone else's, answers\n404.",
        "tags": [
          "exercises"
        ],
        "summary": "Retire an exercise",
        "operationId": "retireExercise",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The exercise's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Retired"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/runs": {
      "get": {
        "description": "Finished or not, newest day first.",
        "tags": [
          "runs"
        ],
        "summary": "List your runs",
        "operationId": "listRuns",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/runListResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "description": "With a template, the run copies it now, so editing the template later never\nreaches the run. Without one, it is an open run that starts empty. A\nwrite-up of a day already gone sends that day as local_date.",
        "tags": [
          "runs"
        ],
        "summary": "Start a run",
        "operationId": "startRun",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/runStartRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/runResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/runs/{id}": {
      "get": {
        "tags": [
          "runs"
        ],
        "summary": "Read a run",
        "operationId": "readRun",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The run's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/runResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "description": "Sets everything but the plan, the template and the finish. A field left out\nor sent as null is cleared. A finished run can still be changed.",
        "tags": [
          "runs"
        ],
        "summary": "Replace a run",
        "operationId": "updateRun",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/runRequest"
            }
          },
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The run's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/runResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "description": "Removes one of your runs with everything logged in it.",
        "tags": [
          "runs"
        ],
        "summary": "Delete a run",
        "operationId": "deleteRun",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The run's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/runs/{id}/finish": {
      "post": {
        "description": "Every step still null becomes skipped. Finishing twice keeps the first\ntime.",
        "tags": [
          "runs"
        ],
        "summary": "Finish a run",
        "operationId": "finishRun",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "ID",
            "description": "The run's id.",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/runResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/session-templates": {
      "get": {
        "description": "The ones the app ships and the ones you wrote, sorted by name. Retired ones\nare left out.",
        "tags": [
          "session-templates"
        ],
        "summary": "List your session templates",
        "operationId": "listSessionTemplates",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/sessionTemplateListResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "description": "A field left out or sent as null is not set. A problem inside a section is\nnamed by its path, such as sections[0].items[2].choice.pick.",
        "tags": [
          "session-templates"
        ],
        "summary": "Create a session template",
        "operationId": "createSessionTemplate",
        "responses": {
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/sessionTemplateResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/session-templates/{id}": {
      "get": {
        "description": "One the app ships or one of yours. A retired one still reads.",
        "tags": [
          "session-templates"
        ],
        "summary": "Read a session template",
        "operationId": "readSessionTemplate",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/sessionTemplateResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "description": "Sets every field of one of your own, sections included. A field left out or\nsent as null is cleared. A shipped session template, or someone else's,\nanswers 404.",
        "tags": [
          "session-templates"
        ],
        "summary": "Replace a session template",
        "operationId": "updateSessionTemplate",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/sessionTemplateResponse"
            }
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/session-templates/{id}/retire": {
      "post": {
        "description": "Takes one of your own out of your list. It still reads by id. Retiring it\ntwice keeps the first date. A shipped session template, or someone else's,\nanswers 404.",
        "tags": [
          "session-templates"
        ],
        "summary": "Retire a session template",
        "operationId": "retireSessionTemplate",
        "responses": {
          "204": {
            "description": "Retired"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          },
          "404": {
            "$ref": "#/responses/notFound"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/tokens": {
      "post": {
        "description": "Each sign-in mints a separate token, so a phone and a laptop hold different\nones and signing out of either leaves the other working.",
        "tags": [
          "tokens"
        ],
        "summary": "Sign in",
        "operationId": "signIn",
        "parameters": [
          {
            "x-go-name": "Body",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/signInRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "$ref": "#/responses/tokenResponse"
          },
          "401": {
            "$ref": "#/responses/invalidCredentials"
          },
          "422": {
            "$ref": "#/responses/validationFailed"
          }
        },
        "security": []
      }
    },
    "/api/v1/tokens/current": {
      "delete": {
        "description": "Deletes the token that authenticated the request. Signing out twice with the\nsame token answers 401 the second time, because the route authenticates\nbefore it deletes.",
        "tags": [
          "tokens"
        ],
        "summary": "Sign out this device",
        "operationId": "signOut",
        "responses": {
          "204": {
            "description": "Signed out"
          },
          "401": {
            "$ref": "#/responses/unauthenticated"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "exerciseRequest": {
      "type": "object",
      "required": [
        "name",
        "kind"
      ],
      "properties": {
        "duration_seconds": {
          "description": "For an open block: how long it runs.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "DurationSeconds",
          "example": 180
        },
        "kind": {
          "description": "Which player screen runs it: reps_and_sets, timed_reps, climbing or open.",
          "type": "string",
          "x-go-name": "Kind",
          "example": "timed_reps"
        },
        "media": {
          "description": "Clips of the exercise, in order. An entry with neither link is dropped.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/mediaBody"
          },
          "x-go-name": "Media"
        },
        "name": {
          "description": "Up to 200 characters.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Repeaters 7:3"
        },
        "notes": {
          "type": "string",
          "x-go-name": "Notes",
          "example": "7 seconds on, 3 off. Stop when the grip goes"
        },
        "per_side": {
          "description": "True when the numbers count for each side, as in 6 reps per side.",
          "type": "boolean",
          "x-go-name": "PerSide"
        },
        "prep_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "PrepSeconds",
          "example": 10
        },
        "rep_rest_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RepRestSeconds",
          "example": 3
        },
        "rep_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "RepSeconds",
          "example": 7
        },
        "reps": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Reps",
          "example": 6
        },
        "set_rest_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "SetRestSeconds",
          "example": 180
        },
        "sets": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Sets",
          "example": 4
        },
        "source": {
          "description": "Who or what the method is known by.",
          "type": "string",
          "x-go-name": "Source",
          "example": "Eva López"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags",
          "example": [
            "hangboard",
            "fingers"
          ]
        }
      },
      "x-go-package": "passion/server/api"
    },
...
    "runListResponse": {
      "type": "object",
      "properties": {
        "runs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/runSummaryBody"
          },
          "x-go-name": "Runs"
        }
      },
      "x-go-package": "passion/server/api"
    },
    "runRequest": {
      "type": "object",
      "required": [
        "name",
        "started_at",
        "local_date"
      ],
      "properties": {
        "elapsed_seconds": {
          "description": "What the session timer showed, in seconds.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "ElapsedSeconds",
          "example": 4200
        },
        "energy": {
          "description": "Energy before the session, 1 to 5.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Energy",
          "example": 4
        },
        "focus": {
          "description": "strength, endurance, technique, projects or general.",
          "type": "string",
          "x-go-name": "Focus",
          "example": "technique"
        },
        "local_date": {
          "description": "The day it counts on, written YYYY-MM-DD.",
          "type": "string",
          "x-go-name": "LocalDate",
          "example": "2026-03-07"
        },
        "name": {
          "description": "Up to 200 characters.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Power"
        },
        "next_focus": {
          "type": "string",
          "x-go-name": "NextFocus",
          "example": "Start the second set less tired"
        },
        "notes": {
          "type": "string",
          "x-go-name": "Notes"
        },
        "place": {
          "description": "Where it happened. A copy of the name, so renaming a gym later changes no\npast run.",
          "type": "string",
          "x-go-name": "Place",
          "example": "The Castle"
        },
        "rpe": {
          "description": "How hard the whole session felt, 1 to 10.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "RPE",
          "example": 7
        },
        "sections": {
          "description": "The working list, in order.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/runSectionBody"
          },
          "x-go-name": "Sections"
        },
        "setting": {
          "description": "indoor or outdoor.",
          "type": "string",
          "x-go-name": "Setting",
          "example": "indoor"
        },
        "sleep": {
          "description": "Sleep quality, 1 to 5.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Sleep",
          "example": 3
        },
        "started_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "StartedAt"
        },
        "went_well": {
          "type": "string",
          "x-go-name": "WentWell",
          "example": "Left hand felt solid"
        }
      },
      "x-go-package": "passion/server/api"
    },
    "runResponse": {
      "type": "object",
      "properties": {
        "elapsed_seconds": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "ElapsedSeconds"
        },
        "energy": {
          "description": "Energy before the session, 1 to 5.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Energy",
          "example": 4
        },
        "finished_at": {
          "description": "null until the run is finished. An unfinished run counts for nothing.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "FinishedAt"
        },
        "focus": {
          "description": "strength, endurance, technique, projects or general.",
          "type": "string",
          "x-go-name": "Focus",
          "example": "technique"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID",
          "example": "01a0bf77-d7e8-76ea-96cc-f09cbca175a3"
        },
        "local_date": {
          "type": "string",
          "x-go-name": "LocalDate",
          "example": "2026-03-07"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "Power"
        },
        "next_focus": {
          "type": "string",
          "x-go-name": "NextFocus",
          "example": "Start the second set less tired"
        },
        "notes": {
          "type": "string",
          "x-go-name": "Notes"
        },
        "place": {
          "type": "string",
          "x-go-name": "Place"
        },
        "plan": {
          "description": "The template's sections as they were when the run started. null for an\nopen run. Nothing changes it.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/sectionBody"
          },
          "x-go-name": "Plan"
        },
        "rpe": {
          "description": "How hard the whole session felt, 1 to 10.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "RPE",
          "example": 7
        },
        "sections": {
          "description": "The working list.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/runSectionBody"
          },
          "x-go-name": "Sections"
        },
        "setting": {
          "description": "indoor or outdoor.",
          "type": "string",
          "x-go-name": "Setting",
          "example": "indoor"
        },
        "sleep": {
          "description": "Sleep quality, 1 to 5.",
          "type": "integer",
          "format": "int64",
          "x-go-name": "Sleep",
          "example": 3
        },
        "started_at": {
          "type": "string",
          "format": "date-time",
          "x-go-name": "StartedAt"
        },
        "template": {
          "description": "The session template it was copied from. null for an open run.",
          "type": "string",
          "x-go-name": "Template"
        },
        "timezone": {
          "description": "The time zone the run's day was worked out in.",
          "type": "string",
          "x-go-name": "Timezone",
          "example": "Europe/London"
        },
        "went_well": {
          "type": "string",
          "x-go-name": "WentWell",
          "example": "Left hand felt solid"
        }
      },
      "x-go-package": "passion/server/api"
    },
...
    "runStartRequest": {
      "type": "object",
      "properties": {
        "local_date": {
          "description": "The day it counts on, written YYYY-MM-DD, for a write-up of a day already\ngone. Defaults to the day of started_at in your time zone.",
          "type": "string",
          "x-go-name": "LocalDate",
          "example": "2026-03-07"
        },
        "name": {
          "description": "Required for an open run, up to 200 characters. A planned run takes its\ntemplate's name.",
          "type": "string",
          "x-go-name": "Name",
          "example": "Open session"
        },
        "started_at": {
          "description": "When it started. Defaults to now.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "StartedAt"
        },
        "template": {
          "description": "A session template to copy now. Leave it out for an open run.",
          "type": "string",
          "x-go-name": "Template",
          "example": "01a0bf77-d7e8-76ea-96cc-f09cbca175a3"
        }
      },
      "x-go-package": "passion/server/api"
    },
...
    "sessionTemplateListResponse": {
      "type": "object",
      "properties": {
        "session_templates": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/sessionTemplateResponse"
          },
          "x-go-name": "SessionTemplates"
        }
      },
      "x-go-package": "passion/server/api"
    },
...
    "sessionTemplateResponse": {
      "type": "object",
      "properties": {
        "color": {
          "type": "string",
          "x-go-name": "Color"
        },
        "id": {
          "type": "string",
          "x-go-name": "ID",
          "example": "01a0bf77-d7e8-76ea-96cc-f09cbca175a3"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name",
          "example": "Power"
        },
        "needs": {
          "type": "string",
          "x-go-name": "Needs"
        },
        "notes": {
          "type": "string",
          "x-go-name": "Notes"
        },
        "retired_at": {
          "description": "When it left the list. A retired session template still reads.",
          "type": "string",
          "format": "date-time",
          "x-go-name": "RetiredAt"
        },
        "sections": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/sectionBody"
          },
          "x-go-name": "Sections"
        },
        "shipped": {
          "description": "True when the app ships it. Nobody can change a shipped session template.",
          "type": "boolean",
          "x-go-name": "Shipped"
        },
        "source": {
          "type": "string",
          "x-go-name": "Source"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Tags"
        }
      },
      "x-go-package": "passion/server/api"
    },
...
    "signUpRequest": {
      "type": "object",
      "required": [
        "email",
        "password",
        "display_name",
        "timezone"
      ],
      "properties": {
        "display_name": {
          "description": "The name shown in the app.",
          "type": "string",
          "x-go-name": "DisplayName",
          "example": "Ada"
        },
        "email": {
          "type": "string",
          "x-go-name": "Email",
          "example": "ada@example.com"
        },
        "password": {
          "description": "Between 8 and 128 characters.",
          "type": "string",
          "x-go-name": "Password",
          "example": "correct horse battery"
        },
        "timezone": {
          "description": "A zone name Postgres knows. It decides which local date a record counts\nas, so a session finished before midnight stays on that day.",
          "type": "string",
          "x-go-name": "Timezone",
          "example": "Europe/Oslo"
        }
      },
      "x-go-package": "passion/server/api"
    },
...
  },
  "responses": {
    "accountResponse": {
      "description": "An account.",
      "schema": {
        "$ref": "#/definitions/accountResponse"
      }
    },
    "exerciseListResponse": {
      "description": "Your library.",
      "schema": {
        "$ref": "#/definitions/exerciseListResponse"
      }
    },
    "exerciseResponse": {
      "description": "An exercise.",
      "schema": {
        "$ref": "#/definitions/exerciseResponse"
      }
    },
    "invalidCredentials": {
      "description": "The address and password do not match an account.",
      "schema": {
        "$ref": "#/definitions/errorBody"
      }
    },
    "notFound": {
      "description": "Nothing you can see or change has that id.",
      "schema": {
        "$ref": "#/definitions/errorBody"
      }
    },
    "signUpResponse": {
      "description": "The account and its first token.",
      "schema": {
        "$ref": "#/definitions/signUpResponse"
      }
    },
...
    "unauthenticated": {
      "description": "No live bearer token.",
      "schema": {
        "$ref": "#/definitions/errorBody"
      }
    },
    "validationFailed": {
      "description": "Something in the body needs fixing.",
      "schema": {
        "$ref": "#/definitions/errorBody"
      }
    }
  },
  "securityDefinitions": {
    "bearer": {
      "description": "The token from a sign-up or sign-in response, sent as \"Bearer \u003ctoken\u003e\". It lasts thirty days, unless the server sets another length, and slides forward on use.",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
...
}